requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==5.1.0
openai==1.3.7
langchain==0.0.335
langchain-community==0.0.1
pinecone-client==3.0.2
docx2txt==0.8
ollama==0.1.6
pydantic==2.5.0
